pyproject.toml
[project]
name = "binance-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Oscar Cortez",email = "oscarkortez@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-binance (>=1.0.34,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
